Move AWS Configuration To Services Directory
