+++ pages/api/mail.ts
@@ -1,5 +1,5 @@
 // Nodemailer with SES transport, more info here: https://nodemailer.com/transports/ses/
-import { transporter } from "aws/index";
+import { transporter } from "services/aws/index";
 import { ResponseObject } from "models/response";
 import type { NextApiRequest, NextApiResponse } from "next";
